pyproject.toml
[tool.poetry]
name = "youtube-notificator"
version = "0.1.0"
description = ""
authors = ["fed0r <wisp47344@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.0"
sqlalchemy = "^2.0.34"
uvicorn = "^0.30.6"
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
dependency-injector = "^4.41.0"
aiogram = "^3.13.0"
asyncpg = "^0.29.0"
aiokafka = "^0.11.0"
requests = "^2.32.3"
google-api-python-client = "^2.149.0"
python-multipart = "^0.0.12"
passlib = "^1.7.4"
pycryptodome = "^3.21.0"
pyjwt = "^2.9.0"
datetime = "^5.5"
bs4 = "^0.0.2"
dateparser = "^1.2.0"
isodate = "^0.7.2"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
